requirements.txt
# For python projects only to set up venv
# Do not use packages outside of safelist. Speak to head of data science and have package added to safe list
# The libraries used by your code should be listed here
# See https://nhsd-git.digital.nhs.uk/data-services/analytics-service/iuod/rap-community-of-practice/-/blob/master/python/project-structure-and-packaging.md

# Python version = 3.12.*

# Data manipulation
numpy==1.26.0
pandas==2.2.2
pyxlsb==1.0.10

# SQL connections
pyodbc==5.1.0
sqlalchemy==2.0.30
ncl_sqlsnippets==1.3.0

# Excel output
#openpyxl==3.0.9

# Testing
#pytest==6.2.5
#pytest-html==3.1.1

# Dependencies of the above packages
#ipykernel==6.9.0
#nbformat==5.1.3
toml==0.10.2
#pathlib2==2.3.6

#Temporary environment
python-dotenv==1.0.1
requests==2.31.0